+++ CHANGELOG.rst
@@ -11,12 +11,16 @@ Added
 - Added a UI button for redeploying an EVC.
 - UNI tag_type are now accepted as string.
 - Event ``kytos/mef_eline.evcs_loaded`` gets published during NApp setup
+- EVCs now listen to ``switch.interface.(link_up|link_down|created|deleted)`` events for activation/deactivation
+- Circuits with a vlan range are supported now. The ranges follows ``list[list[int]]`` format and both UNIs vlan should have the same ranges.
 
 Changed
 =======
+- EVCs will try to maintain their current_path on link status changes
 - UNIs now will use and free tags from ``Interface.available_tags``.
 - UNI tag_type is changed to string from 1, 2 and 3 values to ``"vlan"``, ``"vlan_qinq"`` and ``"mpls"`` respectively.
 - Add ``set_vlan`` only if UNI A vlan and UNI z vlan are different.
+- Updated ``openapi.yml``, ``Tag`` now can accept ``array`` as ``value``.
 
 Deprecated
 ==========